fix(csa-hooks): route post-merge sync output to stderr not stdout (#1070)
gemini-code-assist HIGH finding: `{ ...; } 2>&1` leaked git output to
wrapper stdout, breaking callers that parse `gh --json`. Fix: `{ ...; } >&2`
routes both streams to stderr. Split post-merge sync tests to separate file
to stay under monolith-file limit.
